add check for default image to not delete

+++ app/Http/Controllers/AdminPostController.php
@@ -73,7 +73,9 @@ class AdminPostController extends Controller
 
     public function destroy(Post $post)
     {
-        Storage::delete($post->thumbnail);
+        if(!$post->thumbnail == "thumbnails/default-img.png"){
+            Storage::delete($post->thumbnail);
+        }
         $post->delete();
 
         return back()->with('success', 'Post has been Deleted');
